+++ packages/ui/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from "./useDimensions";
+export * from "./useFontStyle";
 export * from "./useReduceMotion";
 export * from "./useResources";
 export * from "./useScalePress";